requirements.txt
# Langfuse SDK
langfuse>=2.0.0

# LLM Provider SDKs
openai>=1.0.0
# anthropic>=0.25.0
# google-generativeai>=0.3.0

# Framework Integrations (optional)
# langchain>=0.1.0
# langchain-openai>=0.0.5
# llama-index>=0.9.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Data handling
pandas>=2.0.0

# Optional: For advanced examples
# pydantic>=2.0.0
# asyncio
